Restructured code to be easier to navigate
