+++ lib/view/home/home_master.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:temp/components/topnav.dart';
+
+import 'home_mobile.dart';
+import 'home_web.dart';
+
+class HomePageMaster extends StatefulWidget {
+  @override
+  _HomePageMasterState createState() => _HomePageMasterState();
+}
+
+class _HomePageMasterState extends State<HomePageMaster> {
+  bool isLargeScreen;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: TopNav().getAppBar(),
+      drawer: TopNav().getDrawer(),
+      body: OrientationBuilder(
+        builder: (context, orientation) {
+          if (MediaQuery.of(context).size.width > 600) {
+            isLargeScreen = true;
+          } else {
+            isLargeScreen = false;
+          }
+
+          if (isLargeScreen) {
+            return HomePageWeb();
+          } else {
+            return HomePageMobile();
+          }
+        },
+      ),
+    );
+  }
+}
