+++ lib/view/login.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:temp/view/home_master.dart';
-
-import 'home_mobile.dart';
+import 'package:temp/view/home/home_master.dart';
 
 class LoginPage extends StatefulWidget {
   @override
